Self notes on the ERD & Crow's Notation
Pertaining to the utter confusion between notations used in Entity-Relationship modeling concepts and especially the degree v/s cardinality question, I have compiled my notes on most of the concepts used in ERD from authentic and reasonable sources which do not contradict each other. I did this because there are misconceptions over the internet for simple definitions which make it cumbersome to understand. There are innumerable errors both in the teacher's slides as well as articles over Internet in defining terms and not keeping the exact semantics of it while defining subsequent terms. Hope this will do justice to those who are behind the true meaning of everything in between! Also through discovery, I noticed the Crow's notation which makes things really simple to understand. Thanks to Coursera and Patrycja Dybka! Copyrights and licensing reserved